pyproject.toml
[project]
name = "webpush"
version = "1.0.5"
description = "WebPush library for python"
authors = [{ name = "Alexey", email = "delvinru@gmail.com" }]
requires-python = "~=3.10"
readme = "README.md"
license = "MIT"
keywords = ["webpush", "pwa", "pydantic"]
dependencies = [
    "cryptography>=44.0.0",
    "pydantic>=2.10.5",
    "email-validator>=2.2.0",
    "pyjwt>=2.10.1",
]

[project.urls]
Homepage = "https://github.com/delvinru/webpush-py"
Repository = "https://github.com/delvinru/webpush-py"

[project.scripts]
vapid-gen = "webpush.cli:main"

[dependency-groups]
dev = ["pynvim>=0.5.2", "pytest>=8.3.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
